requirements.txt
# Python project dependencies
google-generativeai
python-dotenv
numpy